Add train file.
